Move HashString to cacheutils
